create  createChatClient in WebviewHelper
