+++ clients/vscode/src/chat/WebviewHelper.ts
@@ -30,6 +30,7 @@ import { GitProvider } from "../git/GitProvider";
 import { getLogger } from "../logger";
 import { contributes } from "../../package.json";
 import { parseKeybinding, readUserKeybindingsConfig } from "../util/KeybindingParser";
+import { createClient } from "./chatPanel";
 
 export class WebviewHelper {
   webview?: Webview;
@@ -442,6 +443,119 @@ export class WebviewHelper {
     } else {
       this.displayDisconnectedPage();
     }
+  }
+
+  public createChatClient(webview: Webview) {
+    return createClient(webview, {
+      navigate: async (context: Context, opts?: NavigateOpts) => {
+        if (opts?.openInEditor) {
+          const document = await WebviewHelper.resolveDocument(this.logger, workspace.workspaceFolders, context.filepath);
+          if (!document) {
+            throw new Error(`File not found: ${context.filepath}`);
+          }
+
+          const newEditor = await window.showTextDocument(document, {
+            viewColumn: ViewColumn.Active,
+            preview: false,
+            preserveFocus: true,
+          });
+
+          // Move the cursor to the specified line
+          const start = new Position(Math.max(0, context.range.start - 1), 0);
+          const end = new Position(context.range.end, 0);
+          newEditor.selection = new Selection(start, end);
+          newEditor.revealRange(new Range(start, end), TextEditorRevealType.InCenter);
+
+          return;
+        }
+
+        if (context?.filepath && context?.git_url) {
+          const serverInfo = await this.agent.fetchServerInfo();
+
+          const url = new URL(`${serverInfo.config.endpoint}/files`);
+          const searchParams = new URLSearchParams();
+
+          searchParams.append("redirect_filepath", context.filepath);
+          searchParams.append("redirect_git_url", context.git_url);
+          url.search = searchParams.toString();
+
+          const lineHash = this.formatLineHashForCodeBrowser(context.range);
+          if (lineHash) {
+            url.hash = lineHash;
+          }
+
+          await env.openExternal(Uri.parse(url.toString()));
+        }
+      },
+      refresh: async () => {
+        const serverInfo = await this.agent.fetchServerInfo();
+        await this.displayChatPage(serverInfo.config.endpoint, { force: true });
+        return;
+      },
+      onSubmitMessage: async (msg: string, relevantContext?: Context[]) => {
+        const editor = window.activeTextEditor;
+        const chatMessage: ChatMessage = {
+          message: msg,
+          relevantContext: [],
+        };
+        if (editor) {
+          const fileContext = WebviewHelper.getFileContextFromSelection({ editor, gitProvider: this.gitProvider });
+          if (fileContext)
+            // active selection
+            chatMessage.activeContext = fileContext;
+        }
+        if (relevantContext) {
+          chatMessage.relevantContext = chatMessage.relevantContext?.concat(relevantContext);
+        }
+
+        // FIXME: maybe deduplicate on chatMessage.relevantContext
+        this.sendMessage(chatMessage);
+      },
+      onApplyInEditor: (content: string) => {
+        const editor = window.activeTextEditor;
+        if (editor) {
+          const document = editor.document;
+          const selection = editor.selection;
+
+          // Determine the indentation for the content
+          // The calculation is based solely on the indentation of the first line
+          const lineText = document.lineAt(selection.start.line).text;
+          const match = lineText.match(/^(\s*)/);
+          const indent = match ? match[0] : "";
+
+          // Determine the indentation for the content's first line
+          // Note:
+          // If using spaces, selection.start.character = 1 means 1 space
+          // If using tabs, selection.start.character = 1 means 1 tab
+          const indentUnit = indent[0];
+          const indentAmountForTheFirstLine = Math.max(indent.length - selection.start.character, 0);
+          const indentForTheFirstLine = indentUnit?.repeat(indentAmountForTheFirstLine) || "";
+
+          // Indent the content
+          const indentedContent = indentForTheFirstLine + content.replaceAll("\n", "\n" + indent);
+
+          // Apply into the editor
+          editor.edit((editBuilder) => {
+            editBuilder.replace(selection, indentedContent);
+          });
+        }
+      },
+      onLoaded: () => {
+        setTimeout(() => {
+          this.refreshChatPage();
+        }, 300);
+      },
+      onCopy: (content) => {
+        env.clipboard.writeText(content);
+      },
+      focusOnEditor: () => {
+        const editor = window.activeTextEditor;
+        if (editor) {
+          getLogger().info("Focus back to active editor");
+          commands.executeCommand("workbench.action.focusFirstEditorGroup");
+        }
+      },
+    });
   }
 }
